+++ FlintstonesEntities/BetEntity.cs
@@ -17,14 +17,14 @@ namespace FlintstonesEntities
 
         public string Token { get; set; }
         public int StatusID { get; set; } // 1: Active, 2: Win, 3: Lose, 4: Refund
-        public decimal StakeAmount { get; set; }
+        public double StakeAmount { get; set; }
         public string Market { get; set; }
         public int Selection { get; set; }
-        public decimal SelectionOdd { get; set; }
+        public double SelectionOdd { get; set; }
         public int Duration { get; set; }
         public double CurrentMarketPrice { get; set; }
         public string? Tag { get; set; }
-        public decimal TotalPayout { get { return StakeAmount * SelectionOdd; } }
+        public double TotalPayout { get { return StakeAmount * SelectionOdd; } }
         public DateTime CreatedDate { get; set; }
     }
 }
